Fix ResourceUrl for Token Retrieval in Connect-FabricAccount (#51)
* Fix ResourceUrl for Token Retrieval in Connect-FabricAccount

Updated the ResourceUrl used to retrieve the authentication token in the Connect-FabricAccount function. This change ensures that the correct API endpoint is utilized for obtaining the Fabric token, improving the reliability of the connection process for Pester Help Tests.

Thank you!

* Fixed a bug in New-FabricEventhouse that ran Invoke-RestMethod twice. Removed the ShouldProcess Check, rather than the one outside the ShouldProcess This change simplifies the API request process.

* Add ShouldProcess check for API request in New-FabricEventhouse

This change introduces a ShouldProcess check before making the API request to create a new Eventhouse. This ensures that the operation is confirmed by the user, enhancing safety and control over the action.