Add GitHub Actions workflow for module deployment
This commit introduces a new GitHub Actions workflow to automate the deployment of the module to the PowerShell Gallery. It includes steps for packaging the module, publishing it, and creating a changelog pull request. This enhancement aims to streamline the release process and ensure consistency in deployments for Pester Help Tests.

Thank you!