Today Rob will mainly be creating PRs for deployments so that Jess can approve them (#68)
* Remove permissions from deploy preview workflow for Pester Help Tests

This change simplifies the workflow by removing unnecessary permissions that are not required for the deployment process. Additionally, it corrects the GitHub token secret reference to ensure proper access during the publish step.

Thank you!

* Update comment for ScriptsToProcess section

Clarify the comment for the ScriptsToProcess section in the module manifest. This change enhances readability and understanding of the purpose of this section for future developers.

Thank you!