we wont tag (#69)
* Remove permissions from deploy preview workflow for Pester Help Tests

This change simplifies the workflow by removing unnecessary permissions that are not required for the deployment process. Additionally, it corrects the GitHub token secret reference to ensure proper access during the publish step.

Thank you!

* Update comment for ScriptsToProcess section

Clarify the comment for the ScriptsToProcess section in the module manifest. This change enhances readability and understanding of the purpose of this section for future developers.

Thank you!

* Remove Create_Release_Git_Tag from publish workflow for Pester Help Tests

This change simplifies the publish workflow by removing the Create_Release_Git_Tag step. This adjustment helps streamline the process and ensures that the workflow runs more efficiently.

Thank you!

* Update comment for ScriptsToProcess section for Pester Help Tests

Clarify the purpose of the ScriptsToProcess section in the module manifest. This change enhances readability and understanding for future developers.

Thank you!