So much fun to be had testing deployment actions (#111)
* Reorder Send Changelog PR step in deployment workflow

Moved the Send Changelog PR step to ensure it runs after setting the Git configuration. This change improves the workflow's reliability by ensuring that the necessary Git settings are applied before attempting to create a changelog pull request.

Thank you!

* Remove extra blank line in Set-FabricConfig.ps1

This change cleans up the formatting of the Set-FabricConfig.ps1 file by removing an unnecessary blank line. This improves readability and maintains consistency in the code style.

Thank you!

* Update Git config for automated deployment

Changed the Git user email and name in the deployment workflow to reflect the automated process. This ensures that commits made during the deployment are attributed correctly.

Thank you!