Feature/small updates (#122)
* Fix function name in CHANGELOG and update examples

Corrected the function name from Get-SqlDatabase to Get-FabricSqlDatabase in the CHANGELOG.
Updated examples in the Get-FabricSQLDatabase function to improve clarity and consistency.
Thank you!

* Update Get-FabricSQLDatabase to improve parameter handling

Refactor the function to return warnings instead of throwing errors when both SQLDatabaseName and SQLDatabaseId are provided. Update the synopsis and description for clarity. Ensure the return value is correctly handled in the End block.

Thank you!

* Update Get-FabricSQLDatabase documentation and URI

Clarify the function's synopsis and description to enhance understanding.
Corrected the spelling of "examples" and updated the URI to follow the correct casing for SQL databases.

Thank you!

* Update example in Get-FabricSQLDatabase documentation

Changed the example command to use quotes around the workspace name for consistency and clarity. This helps users understand the correct syntax when using the function.

Thank you!

* Update example in Get-FabricSQLDatabase documentation

Changed the workspace name in the example to 'MsLearn-dev' because thats important

Thank you!

* Refactor Get-FabricSQLDatabase to improve parameter handling

This change restructures the function to ensure that the SQLDatabaseName and SQLDatabaseID parameters cannot be used together, enhancing the clarity of parameter usage. The process block has been properly defined to maintain code organization and readability.

Thank you!

* Refactor Get-FabricSQLDatabase for improved readability

Reorganize function structure and enhance code clarity by adjusting the formatting of the function definition and control structures. This change aims to make the code more maintainable and easier to understand.

Thank you!

* Initialize return array in Get-FabricSQLDatabase

This change introduces an initialization of the return array at the beginning of the Get-FabricSQLDatabase function. This will ensure that the function has a defined return structure, improving its reliability and maintainability.

Thank you!

* Add check for response before updating return array

This change ensures that the return array is only updated if there is a valid response from the SQLDatabase filtering. This prevents potential errors when no matching databases are found.

Thank you!