Update Get-FabricWorkspaceUser to handle null responses
should never be required but use lessons learned from the database code
Enhance the Get-FabricWorkspaceUser function to ensure that
it properly checks for null responses from the PowerBI API
before adding users to the return value. This change improves
the robustness of the function and prevents potential errors
when no users are returned.

Thank you!