Enable Get-FabricWorkspaceUser to accept pipeline input (#128)
* Rename function, improve examples and add alias

* Refactor Get-FabricWorkspaceUserto enable processing via the pipeline  #75

Updated the Get-FabricWorkspaceUser function to enhance readability and maintainability. Added additional examples to the documentation for better user guidance. Improved the process block to handle multiple workspaces and return user details more effectively.

Thank you!

* Add unit tests for Get-FabricWorkspaceUser function

This commit introduces a comprehensive set of unit tests for the Get-FabricWorkspaceUser function. The tests validate parameter presence, alias functionality, and ensure correct behavior when handling multiple workspaces. This enhancement improves code reliability and maintainability.

Thank you!

* Add unit tests and update Get-FabricWorkspaceUser

Added unit tests for the `Get-FabricWorkspaceUser` function to ensure it operates correctly with multiple workspaces, both in the pipeline and as parameters. Updated the function to support pipeline input for `WorkspaceId` and `WorkspaceName`, and renamed `Get-FabricWorkspaceUsers` to its singular form for consistency.

Thank you!

* Refactor test for Get-FabricWorkspaceUser to use script block

Updated the test case to wrap the command in a script block for better readability and consistency. This change enhances the clarity of the test and aligns with best practices for PowerShell testing.

Thank you!

* Update Get-FabricWorkspaceUser to handle null responses

should never be required but use lessons learned from the database code
Enhance the Get-FabricWorkspaceUser function to ensure that
it properly checks for null responses from the PowerBI API
before adding users to the return value. This change improves
the robustness of the function and prevents potential errors
when no users are returned.

Thank you!

* Remove comments