Update all id parameters to use guids instead of strings as datatypes closes #125 (#126)
* Change WorkspaceId and LakehouseId parameters to GUID Replace type for all `Id` input parameters from [String] to [Guid] #125

Updated parameter types for WorkspaceId and LakehouseId to GUID in multiple scripts to ensure better validation and consistency across the codebase. This change enhances the robustness of the functions by enforcing the correct data type.

Thank you!

* Refactor parameter types in PowerShell scripts to use [guid] for WorkspaceId, ReflexId, ReportId, and other identifiers

- Updated parameter types from [string] to [guid] in multiple scripts across the Public/Reflex, Public/Report, Public/Restore Points, Public/SQL Database, Public/Spark, Public/Tenant, Public/Users, Public/Utils, Public/Warehouse, and Public/Workspace directories.
- This change enhances type safety and ensures that only valid GUIDs are accepted for these parameters.

* Update parameter types to GUID in CHANGELOG and script

Updated the `WorkspaceId`, `CapacitiesIds`, and various other parameters to use the GUID datatype for consistency and improved type safety. This change enhances the clarity and reliability of the code.

Thank you!

* Fix syntax error in Export-FabricItem function

Corrected the conditional check for itemID to ensure proper functionality. This change enhances the reliability of the function when exporting items from a Fabric workspace.

Thank you!

* Add tests for parameter types in functions

This commit introduces a new test suite to verify that function parameters ending with 'id' are of type System.Guid. This ensures consistency and correctness in parameter types across the module.

Thank you!

* So that the parameter data type is correct

Updated the WorkspaceId parameter type from string to GUID for better validation and to ensure correct input format. This change enhances the function's robustness and prevents potential errors.

Thank you!